requirements.txt
black
dj-database-url
django
django-cors-headers
django-environ
django-extensions
django-filter
django-localflavor
django-model-utils
django-split-settings
django-storages
django-webpack-loader==1.7.0
djangorestframework
python-dotenv
environ
factory_boy
flake8
getconf
mysqlclient
pre-commit
pytest-cov
pytest-django
pytest-env
pyjwt[crypto]
pytest-lazy-fixture
pytest-watch
pytz
responses
split_settings
jours-feries-france